Initial installation script setup

+++ bootstrap.sh
@@ -1,10 +1,21 @@
 #!/usr/bin/env bash
 # Script to further automate the download and installation of a brand new
-# system. It essentially gets everything ready for "building" the system
-# configurations using "tmpl".
+# system. It essentially builds the appropriate system installation script using
+# "tmpl". The specified template configuration file is relative to the "tmpl"
+# subdirectory of the repository.
 
 set -e
 trap "exit 100" INT
+
+export TMPL_BLOCK_END_STR='%}'
+export TMPL_BLOCK_START_STR='{%'
+export TMPL_COMMENT_END_STR='#}'
+export TMPL_COMMENT_START_STR='{#'
+export TMPL_LOG_FILE='mkdot.log'
+export TMPL_LOG_LEVEL='info'
+export TMPL_LOG_MODE='overwrite'
+export TMPL_VAR_END_STR='}}'
+export TMPL_VAR_START_STR='{{'
 
 # ---------- Configuration ----------
 dotfiles_url='https://github.com/HarrisonTotty/dotfiles/archive/tmpl.zip'
@@ -13,11 +24,17 @@ tmpl_url='https://raw.githubusercontent.com/HarrisonTotty/tmpl/master/tmpl.py'
 
 cd "$HOME"
 
+if [ "$#" -ne 1 ]; then
+    echo "USAGE: ./bootstrap.sh <template conf>"
+    echo "EXAMPLE: ./bootstrap.sh personal-laptop.yaml"
+    exit 0
+fi
+
 if [ -f /usr/local/bin/tmpl ]; then
     rm -f /usr/local/bin/tmpl
 fi
 
-wget "$tmpl_url" -O /usr/local/bin/tmpl && chmod +x /usr/local/bin/tmpl
+wget -q "$tmpl_url" -O /usr/local/bin/tmpl && chmod +x /usr/local/bin/tmpl
 
 pacman -Sy python-jinja python-yaml unzip --noconfirm
 
@@ -25,7 +42,18 @@ if [ -d "$HOME/dotfiles" ]; then
     rm -rf "$HOME/dotfiles"
 fi
 
-wget "$dotfiles_url" -O dotfiles.zip \
+wget -q "$dotfiles_url" -O dotfiles.zip \
     && unzip dotfiles.zip \
     && rm -f dotfiles.zip \
     && mv dotfiles-* dotfiles
+
+if [ -f "$HOME/install-arch.sh" ]; then
+    rm -f "$HOME/install-arch.sh"
+fi
+
+conf_path="$HOME/dotfiles/tmpl/$1"
+script_path="$HOME/dotfiles/src/scripts/install-arch.sh"
+
+cat "$script_path" | tmpl "$conf_path" --stdin > "$HOME/install-arch.sh"
+
+chmod +x "$HOME/install-arch.sh"
